+++ examples/agentic-a2ui-app/CHANGELOG.md
@@ -1,5 +1,15 @@
 # prometheus-agentic-a2ui-example
 
+## 0.1.2
+
+### Patch Changes
+
+- Updated dependencies
+  - @prometheus-ags/entity-graph-core@3.0.3
+  - @prometheus-ags/prometheus-entity-management@3.0.3
+  - @prometheus-ags/a2ui-react@3.0.3
+  - @prometheus-ags/entity-graph-a2a@3.0.3
+
 ## 0.1.1
 
 ### Patch Changes