+++ packages/a2ui-react/CHANGELOG.md
@@ -1,5 +1,12 @@
 # @prometheus-ags/a2ui-react changelog
 
+## 3.0.3
+
+### Patch Changes
+
+- Updated dependencies
+  - @prometheus-ags/entity-graph-core@3.0.3
+
 ## 3.0.0
 
 ### Patch Changes